main_backend/requirements.txt
Django==4.1.3
gunicorn==20.1.0
requests==2.28.1
django-configurations==2.4
django-environ==0.9.0
psycopg2-binary==2.9.5
celery==5.2.7
django-celery-results==2.4.0
redis>=4.3.4
django-celery-beat==2.4.0